extracted mod data to new benchmarking/mock/tests files
