+++ polkadot/runtime/common/src/slots/mod.rs
@@ -497,653 +497,11 @@ impl<T: Config> Leaser<BlockNumberFor<T>> for Pallet<T> {
 	}
 }
 
-/// tests for this pallet
 #[cfg(test)]
-mod tests {
-	use super::*;
+mod mock;
 
-	use crate::{mock::TestRegistrar, slots};
-	use frame_support::{assert_noop, assert_ok, derive_impl, parameter_types};
-	use frame_system::EnsureRoot;
-	use pallet_balances;
-	use polkadot_primitives::BlockNumber;
-	use polkadot_primitives_test_helpers::{dummy_head_data, dummy_validation_code};
-	use sp_core::H256;
-	use sp_runtime::{
-		traits::{BlakeTwo256, IdentityLookup},
-		BuildStorage,
-	};
-
-	type Block = frame_system::mocking::MockBlockU32<Test>;
-
-	frame_support::construct_runtime!(
-		pub enum Test
-		{
-			System: frame_system,
-			Balances: pallet_balances,
-			Slots: slots,
-		}
-	);
-
-	#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-	impl frame_system::Config for Test {
-		type BaseCallFilter = frame_support::traits::Everything;
-		type BlockWeights = ();
-		type BlockLength = ();
-		type RuntimeOrigin = RuntimeOrigin;
-		type RuntimeCall = RuntimeCall;
-		type Nonce = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Block = Block;
-		type RuntimeEvent = RuntimeEvent;
-		type DbWeight = ();
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = pallet_balances::AccountData<u64>;
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = ();
-		type OnSetCode = ();
-		type MaxConsumers = frame_support::traits::ConstU32<16>;
-	}
-
-	#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-	impl pallet_balances::Config for Test {
-		type AccountStore = System;
-	}
-
-	parameter_types! {
-		pub const LeasePeriod: BlockNumber = 10;
-		pub static LeaseOffset: BlockNumber = 0;
-		pub const ParaDeposit: u64 = 1;
-	}
-
-	impl Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type Currency = Balances;
-		type Registrar = TestRegistrar<Test>;
-		type LeasePeriod = LeasePeriod;
-		type LeaseOffset = LeaseOffset;
-		type ForceOrigin = EnsureRoot<Self::AccountId>;
-		type WeightInfo = crate::slots::TestWeightInfo;
-	}
-
-	// This function basically just builds a genesis storage key/value store according to
-	// our desired mock up.
-	pub fn new_test_ext() -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::<Test>::default().build_storage().unwrap();
-		pallet_balances::GenesisConfig::<Test> {
-			balances: vec![(1, 10), (2, 20), (3, 30), (4, 40), (5, 50), (6, 60)],
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-		t.into()
-	}
-
-	fn run_to_block(n: BlockNumber) {
-		while System::block_number() < n {
-			Slots::on_finalize(System::block_number());
-			Balances::on_finalize(System::block_number());
-			System::on_finalize(System::block_number());
-			System::set_block_number(System::block_number() + 1);
-			System::on_initialize(System::block_number());
-			Balances::on_initialize(System::block_number());
-			Slots::on_initialize(System::block_number());
-		}
-	}
-
-	#[test]
-	fn basic_setup_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-			assert_eq!(Slots::lease_period_length(), (10, 0));
-			let now = System::block_number();
-			assert_eq!(Slots::lease_period_index(now).unwrap().0, 0);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-
-			run_to_block(10);
-			let now = System::block_number();
-			assert_eq!(Slots::lease_period_index(now).unwrap().0, 1);
-		});
-	}
-
-	#[test]
-	fn lease_lifecycle_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			assert_ok!(Slots::lease_out(1.into(), &1, 1, 1, 1));
-			assert_eq!(Slots::deposit_held(1.into(), &1), 1);
-			assert_eq!(Balances::reserved_balance(1), 1);
-
-			run_to_block(19);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 1);
-			assert_eq!(Balances::reserved_balance(1), 1);
-
-			run_to_block(20);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-
-			assert_eq!(
-				TestRegistrar::<Test>::operations(),
-				vec![(1.into(), 10, true), (1.into(), 20, false),]
-			);
-		});
-	}
-
-	#[test]
-	fn lease_interrupted_lifecycle_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			assert_ok!(Slots::lease_out(1.into(), &1, 6, 1, 1));
-			assert_ok!(Slots::lease_out(1.into(), &1, 4, 3, 1));
-
-			run_to_block(19);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			run_to_block(20);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 4);
-			assert_eq!(Balances::reserved_balance(1), 4);
-
-			run_to_block(39);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 4);
-			assert_eq!(Balances::reserved_balance(1), 4);
-
-			run_to_block(40);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-
-			assert_eq!(
-				TestRegistrar::<Test>::operations(),
-				vec![
-					(1.into(), 10, true),
-					(1.into(), 20, false),
-					(1.into(), 30, true),
-					(1.into(), 40, false),
-				]
-			);
-		});
-	}
-
-	#[test]
-	fn lease_relayed_lifecycle_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			assert!(Slots::lease_out(1.into(), &1, 6, 1, 1).is_ok());
-			assert!(Slots::lease_out(1.into(), &2, 4, 2, 1).is_ok());
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-			assert_eq!(Slots::deposit_held(1.into(), &2), 4);
-			assert_eq!(Balances::reserved_balance(2), 4);
-
-			run_to_block(19);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-			assert_eq!(Slots::deposit_held(1.into(), &2), 4);
-			assert_eq!(Balances::reserved_balance(2), 4);
-
-			run_to_block(20);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-			assert_eq!(Slots::deposit_held(1.into(), &2), 4);
-			assert_eq!(Balances::reserved_balance(2), 4);
-
-			run_to_block(29);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-			assert_eq!(Slots::deposit_held(1.into(), &2), 4);
-			assert_eq!(Balances::reserved_balance(2), 4);
-
-			run_to_block(30);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-			assert_eq!(Slots::deposit_held(1.into(), &2), 0);
-			assert_eq!(Balances::reserved_balance(2), 0);
-
-			assert_eq!(
-				TestRegistrar::<Test>::operations(),
-				vec![(1.into(), 10, true), (1.into(), 30, false),]
-			);
-		});
-	}
-
-	#[test]
-	fn lease_deposit_increase_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			assert!(Slots::lease_out(1.into(), &1, 4, 1, 1).is_ok());
-			assert_eq!(Slots::deposit_held(1.into(), &1), 4);
-			assert_eq!(Balances::reserved_balance(1), 4);
-
-			assert!(Slots::lease_out(1.into(), &1, 6, 2, 1).is_ok());
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			run_to_block(29);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			run_to_block(30);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-
-			assert_eq!(
-				TestRegistrar::<Test>::operations(),
-				vec![(1.into(), 10, true), (1.into(), 30, false),]
-			);
-		});
-	}
-
-	#[test]
-	fn lease_deposit_decrease_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			assert!(Slots::lease_out(1.into(), &1, 6, 1, 1).is_ok());
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			assert!(Slots::lease_out(1.into(), &1, 4, 2, 1).is_ok());
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			run_to_block(19);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 6);
-			assert_eq!(Balances::reserved_balance(1), 6);
-
-			run_to_block(20);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 4);
-			assert_eq!(Balances::reserved_balance(1), 4);
-
-			run_to_block(29);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 4);
-			assert_eq!(Balances::reserved_balance(1), 4);
-
-			run_to_block(30);
-			assert_eq!(Slots::deposit_held(1.into(), &1), 0);
-			assert_eq!(Balances::reserved_balance(1), 0);
-
-			assert_eq!(
-				TestRegistrar::<Test>::operations(),
-				vec![(1.into(), 10, true), (1.into(), 30, false),]
-			);
-		});
-	}
-
-	#[test]
-	fn clear_all_leases_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			let max_num = 5u32;
-
-			// max_num different people are reserved for leases to Para ID 1
-			for i in 1u32..=max_num {
-				let j: u64 = i.into();
-				assert_ok!(Slots::lease_out(1.into(), &j, j * 10 - 1, i * i, i));
-				assert_eq!(Slots::deposit_held(1.into(), &j), j * 10 - 1);
-				assert_eq!(Balances::reserved_balance(j), j * 10 - 1);
-			}
-
-			assert_ok!(Slots::clear_all_leases(RuntimeOrigin::root(), 1.into()));
-
-			// Balances cleaned up correctly
-			for i in 1u32..=max_num {
-				let j: u64 = i.into();
-				assert_eq!(Slots::deposit_held(1.into(), &j), 0);
-				assert_eq!(Balances::reserved_balance(j), 0);
-			}
-
-			// Leases is empty.
-			assert!(Leases::<Test>::get(ParaId::from(1_u32)).is_empty());
-		});
-	}
-
-	#[test]
-	fn lease_out_current_lease_period() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(2_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			run_to_block(20);
-			let now = System::block_number();
-			assert_eq!(Slots::lease_period_index(now).unwrap().0, 2);
-			// Can't lease from the past
-			assert!(Slots::lease_out(1.into(), &1, 1, 1, 1).is_err());
-			// Lease in the current period triggers onboarding
-			assert_ok!(Slots::lease_out(1.into(), &1, 1, 2, 1));
-			// Lease in the future doesn't
-			assert_ok!(Slots::lease_out(2.into(), &1, 1, 3, 1));
-
-			assert_eq!(TestRegistrar::<Test>::operations(), vec![(1.into(), 20, true),]);
-		});
-	}
-
-	#[test]
-	fn trigger_onboard_works() {
-		new_test_ext().execute_with(|| {
-			run_to_block(1);
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(1_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(2_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-			assert_ok!(TestRegistrar::<Test>::register(
-				1,
-				ParaId::from(3_u32),
-				dummy_head_data(),
-				dummy_validation_code()
-			));
-
-			// We will directly manipulate leases to emulate some kind of failure in the system.
-			// Para 1 will have no leases
-			// Para 2 will have a lease period in the current index
-			Leases::<Test>::insert(ParaId::from(2_u32), vec![Some((0, 0))]);
-			// Para 3 will have a lease period in a future index
-			Leases::<Test>::insert(ParaId::from(3_u32), vec![None, None, Some((0, 0))]);
-
-			// Para 1 should fail cause they don't have any leases
-			assert_noop!(
-				Slots::trigger_onboard(RuntimeOrigin::signed(1), 1.into()),
-				Error::<Test>::ParaNotOnboarding
-			);
-
-			// Para 2 should succeed
-			assert_ok!(Slots::trigger_onboard(RuntimeOrigin::signed(1), 2.into()));
-
-			// Para 3 should fail cause their lease is in the future
-			assert_noop!(
-				Slots::trigger_onboard(RuntimeOrigin::signed(1), 3.into()),
-				Error::<Test>::ParaNotOnboarding
-			);
-
-			// Trying Para 2 again should fail cause they are not currently an on-demand parachain
-			assert!(Slots::trigger_onboard(RuntimeOrigin::signed(1), 2.into()).is_err());
-
-			assert_eq!(TestRegistrar::<Test>::operations(), vec![(2.into(), 1, true),]);
-		});
-	}
-
-	#[test]
-	fn lease_period_offset_works() {
-		new_test_ext().execute_with(|| {
-			let (lpl, offset) = Slots::lease_period_length();
-			assert_eq!(offset, 0);
-			assert_eq!(Slots::lease_period_index(0), Some((0, true)));
-			assert_eq!(Slots::lease_period_index(1), Some((0, false)));
-			assert_eq!(Slots::lease_period_index(lpl - 1), Some((0, false)));
-			assert_eq!(Slots::lease_period_index(lpl), Some((1, true)));
-			assert_eq!(Slots::lease_period_index(lpl + 1), Some((1, false)));
-			assert_eq!(Slots::lease_period_index(2 * lpl - 1), Some((1, false)));
-			assert_eq!(Slots::lease_period_index(2 * lpl), Some((2, true)));
-			assert_eq!(Slots::lease_period_index(2 * lpl + 1), Some((2, false)));
-
-			// Lease period is 10, and we add an offset of 5.
-			LeaseOffset::set(5);
-			let (lpl, offset) = Slots::lease_period_length();
-			assert_eq!(offset, 5);
-			assert_eq!(Slots::lease_period_index(0), None);
-			assert_eq!(Slots::lease_period_index(1), None);
-			assert_eq!(Slots::lease_period_index(offset), Some((0, true)));
-			assert_eq!(Slots::lease_period_index(lpl), Some((0, false)));
-			assert_eq!(Slots::lease_period_index(lpl - 1 + offset), Some((0, false)));
-			assert_eq!(Slots::lease_period_index(lpl + offset), Some((1, true)));
-			assert_eq!(Slots::lease_period_index(lpl + offset + 1), Some((1, false)));
-			assert_eq!(Slots::lease_period_index(2 * lpl - 1 + offset), Some((1, false)));
-			assert_eq!(Slots::lease_period_index(2 * lpl + offset), Some((2, true)));
-			assert_eq!(Slots::lease_period_index(2 * lpl + offset + 1), Some((2, false)));
-		});
-	}
-}
+#[cfg(test)]
+mod tests;
 
 #[cfg(feature = "runtime-benchmarks")]
-mod benchmarking {
-	use super::*;
-	use frame_support::assert_ok;
-	use frame_system::RawOrigin;
-	use polkadot_runtime_parachains::paras;
-	use sp_runtime::traits::{Bounded, One};
-
-	use frame_benchmarking::{account, benchmarks, whitelisted_caller, BenchmarkError};
-
-	use crate::slots::Pallet as Slots;
-
-	fn assert_last_event<T: Config>(generic_event: <T as Config>::RuntimeEvent) {
-		let events = frame_system::Pallet::<T>::events();
-		let system_event: <T as frame_system::Config>::RuntimeEvent = generic_event.into();
-		// compare to the last event record
-		let frame_system::EventRecord { event, .. } = &events[events.len() - 1];
-		assert_eq!(event, &system_event);
-	}
-
-	// Registers a parathread (on-demand parachain)
-	fn register_a_parathread<T: Config + paras::Config>(i: u32) -> (ParaId, T::AccountId) {
-		let para = ParaId::from(i);
-		let leaser: T::AccountId = account("leaser", i, 0);
-		T::Currency::make_free_balance_be(&leaser, BalanceOf::<T>::max_value());
-		let worst_head_data = T::Registrar::worst_head_data();
-		let worst_validation_code = T::Registrar::worst_validation_code();
-
-		assert_ok!(T::Registrar::register(
-			leaser.clone(),
-			para,
-			worst_head_data,
-			worst_validation_code.clone(),
-		));
-		assert_ok!(paras::Pallet::<T>::add_trusted_validation_code(
-			frame_system::Origin::<T>::Root.into(),
-			worst_validation_code,
-		));
-
-		T::Registrar::execute_pending_transitions();
-
-		(para, leaser)
-	}
-
-	benchmarks! {
-		where_clause { where T: paras::Config }
-
-		force_lease {
-			// If there is an offset, we need to be on that block to be able to do lease things.
-			frame_system::Pallet::<T>::set_block_number(T::LeaseOffset::get() + One::one());
-			let para = ParaId::from(1337);
-			let leaser: T::AccountId = account("leaser", 0, 0);
-			T::Currency::make_free_balance_be(&leaser, BalanceOf::<T>::max_value());
-			let amount = T::Currency::minimum_balance();
-			let period_begin = 69u32.into();
-			let period_count = 3u32.into();
-			let origin =
-				T::ForceOrigin::try_successful_origin().map_err(|_| BenchmarkError::Weightless)?;
-		}: _<T::RuntimeOrigin>(origin, para, leaser.clone(), amount, period_begin, period_count)
-		verify {
-			assert_last_event::<T>(Event::<T>::Leased {
-				para_id: para,
-				leaser, period_begin,
-				period_count,
-				extra_reserved: amount,
-				total_amount: amount,
-			}.into());
-		}
-
-		// Worst case scenario, T on-demand parachains onboard, and C lease holding parachains offboard.
-		manage_lease_period_start {
-			// Assume reasonable maximum of 100 paras at any time
-			let c in 0 .. 100;
-			let t in 0 .. 100;
-
-			let period_begin = 1u32.into();
-			let period_count = 4u32.into();
-
-			// If there is an offset, we need to be on that block to be able to do lease things.
-			frame_system::Pallet::<T>::set_block_number(T::LeaseOffset::get() + One::one());
-
-			// Make T parathreads (on-demand parachains)
-			let paras_info = (0..t).map(|i| {
-				register_a_parathread::<T>(i)
-			}).collect::<Vec<_>>();
-
-			T::Registrar::execute_pending_transitions();
-
-			// T on-demand parachains are upgrading to lease holding parachains
-			for (para, leaser) in paras_info {
-				let amount = T::Currency::minimum_balance();
-				let origin = T::ForceOrigin::try_successful_origin()
-					.expect("ForceOrigin has no successful origin required for the benchmark");
-				Slots::<T>::force_lease(origin, para, leaser, amount, period_begin, period_count)?;
-			}
-
-			T::Registrar::execute_pending_transitions();
-
-			// C lease holding parachains are downgrading to on-demand parachains
-			for i in 200 .. 200 + c {
-				let (para, leaser) = register_a_parathread::<T>(i);
-				T::Registrar::make_parachain(para)?;
-			}
-
-			T::Registrar::execute_pending_transitions();
-
-			for i in 0 .. t {
-				assert!(T::Registrar::is_parathread(ParaId::from(i)));
-			}
-
-			for i in 200 .. 200 + c {
-				assert!(T::Registrar::is_parachain(ParaId::from(i)));
-			}
-		}: {
-				Slots::<T>::manage_lease_period_start(period_begin);
-		} verify {
-			// All paras should have switched.
-			T::Registrar::execute_pending_transitions();
-			for i in 0 .. t {
-				assert!(T::Registrar::is_parachain(ParaId::from(i)));
-			}
-			for i in 200 .. 200 + c {
-				assert!(T::Registrar::is_parathread(ParaId::from(i)));
-			}
-		}
-
-		// Assume that at most 8 people have deposits for leases on a parachain.
-		// This would cover at least 4 years of leases in the worst case scenario.
-		clear_all_leases {
-			let max_people = 8;
-			let (para, _) = register_a_parathread::<T>(1);
-
-			// If there is an offset, we need to be on that block to be able to do lease things.
-			frame_system::Pallet::<T>::set_block_number(T::LeaseOffset::get() + One::one());
-
-			for i in 0 .. max_people {
-				let leaser = account("lease_deposit", i, 0);
-				let amount = T::Currency::minimum_balance();
-				T::Currency::make_free_balance_be(&leaser, BalanceOf::<T>::max_value());
-
-				// Average slot has 4 lease periods.
-				let period_count: LeasePeriodOf<T> = 4u32.into();
-				let period_begin = period_count * i.into();
-				let origin = T::ForceOrigin::try_successful_origin()
-					.expect("ForceOrigin has no successful origin required for the benchmark");
-				Slots::<T>::force_lease(origin, para, leaser, amount, period_begin, period_count)?;
-			}
-
-			for i in 0 .. max_people {
-				let leaser = account("lease_deposit", i, 0);
-				assert_eq!(T::Currency::reserved_balance(&leaser), T::Currency::minimum_balance());
-			}
-
-			let origin =
-				T::ForceOrigin::try_successful_origin().map_err(|_| BenchmarkError::Weightless)?;
-		}: _<T::RuntimeOrigin>(origin, para)
-		verify {
-			for i in 0 .. max_people {
-				let leaser = account("lease_deposit", i, 0);
-				assert_eq!(T::Currency::reserved_balance(&leaser), 0u32.into());
-			}
-		}
-
-		trigger_onboard {
-			// get a parachain into a bad state where they did not onboard
-			let (para, _) = register_a_parathread::<T>(1);
-			Leases::<T>::insert(para, vec![Some((account::<T::AccountId>("lease_insert", 0, 0), BalanceOf::<T>::default()))]);
-			assert!(T::Registrar::is_parathread(para));
-			let caller = whitelisted_caller();
-		}: _(RawOrigin::Signed(caller), para)
-		verify {
-			T::Registrar::execute_pending_transitions();
-			assert!(T::Registrar::is_parachain(para));
-		}
-
-		impl_benchmark_test_suite!(
-			Slots,
-			crate::integration_tests::new_test_ext(),
-			crate::integration_tests::Test,
-		);
-	}
-}
+mod benchmarking;
